edge: 修复 CI 暴露的问题（gitleaks 夹具放行 + 测试断言 + int64 转换）
- evidence_test.go：redactor 假 PAT 恢复为精确 36 位夹具值，.gitleaks.toml 正则 allowlist 精确放行（默认规则保留）
- connector_test.go：错误响应 round-trip 断言修正（OK=false 时校验 Error 保留）
- hermes.go：profiles 计数 int64 转换，修复编译错误
- 本地 Go 1.22 工具链验证：edge/agent 与 connectors/hermes build+vet+test 全绿

+++ connectors/hermes/hermes.go
@@ -253,7 +253,7 @@ func collectOp(plan protocol.ScanPlan) (protocol.EvidenceBatch, error) {
 	rootBase := strings.TrimSuffix(strings.TrimSpace(protocol.ExpandHome(sc.Roots[0])), "/*")
 
 	for _, dir := range dirs {
-		if profiles >= limits.MaxFiles {
+		if int64(profiles) >= limits.MaxFiles {
 			batch.Truncated = true
 			break
 		}
